examples/integrations/gcp/src/jobs/etl/images/bounding_box/requirements.txt
google-cloud-storage
# Headless version so we don't need rendering libs
opencv-python-headless
labelbox[data]
PILLOW
